moving CategoryLabel to primitives folder
