+++ src/@primitives/CategoryLabel/CategoryLabel.stories.js
